Mediator for client scanner
