+++ keycloak_scanner/scanners/clients_scanner.py
@@ -3,46 +3,15 @@ from typing import List
 from requests import HTTPError
 
 from keycloak_scanner.logging.vuln_flag import VulnFlag
-from keycloak_scanner.scanners.json_result import JsonResult
-from keycloak_scanner.scanners.realm_scanner import Realms
 from keycloak_scanner.scanners.scanner import Scanner
-from keycloak_scanner.scanners.scanner_pieces import Need2
-from keycloak_scanner.scanners.types import Realm, WellKnownDict
+from keycloak_scanner.scanners.types import Realm, WellKnown, clientType, Client, ClientConfig
 
 
-class ClientConfig(JsonResult):
-    pass
-
-
-class Client:
-
-    def __init__(self, name: str, url: str, auth_endpoint: str = None, client_registration: ClientConfig = None):
-        self.name = name
-        self.url = url
-        self.auth_endpoint = auth_endpoint
-        self.client_registration = client_registration
-
-    def __repr__(self):
-        return f"Client({repr(self.name)}, {repr(self.url)}, {repr(self.auth_endpoint)}, {repr(self.client_registration)})"
-
-    def __eq__(self, other):
-        if isinstance(other, Client):
-            return self.name == other.name \
-                   and self.url == other.url \
-                   and self.auth_endpoint == other.auth_endpoint \
-                   and self.client_registration == other.client_registration
-        return NotImplemented
-
-
-class Clients(List[Client]):
-    pass
-
-
-class ClientScanner(Scanner[Clients], Need2[Realms, WellKnownDict]):
+class ClientScanner(Scanner[Client]):
 
     def __init__(self, clients: List[str], **kwargs):
         self.clients = clients
-        super().__init__(**kwargs)
+        super().__init__(result_type=clientType, **kwargs)
 
     def has_endpoint(self, realm: Realm, client_name: str) -> str:
         url = f'{super().base_url()}/auth/realms/{realm.name}/{client_name}'
@@ -55,22 +24,21 @@ class ClientScanner(Scanner[Clients], Need2[Realms, WellKnownDict]):
         except Exception as e:
             super().info(f'[ClientScanner]: {e}')
 
-    def perform(self, realms: Realms, well_known_dict: WellKnownDict, **kwargs) -> (Clients, VulnFlag):
+    def perform(self, realm: Realm, well_known: WellKnown, **kwargs) -> (List[Client], VulnFlag):
 
-        result: Clients = Clients()
+        result: List[Client] = []
 
-        for realm in realms:
-            for client_name in self.clients:
+        for client_name in self.clients:
 
-                url = self.has_endpoint(realm, client_name)
+            url = self.has_endpoint(realm, client_name)
 
-                # TODO: auth endpoint in other scanner ?
-                auth_url = self.has_auth_endpoint(client_name, realm, well_known_dict)
+            # TODO: auth endpoint in other scanner ?
+            auth_url = self.has_auth_endpoint(client_name, realm, well_known)
 
-                registration = self.get_registration(client_name, realm)
+            registration = self.get_registration(client_name, realm)
 
-                if auth_url is not None or url is not None or registration is not None:
-                    result.append(Client(name=client_name, auth_endpoint=auth_url, url=url, client_registration=registration))
+            if auth_url is not None or url is not None or registration is not None:
+                result.append(Client(name=client_name, auth_endpoint=auth_url, url=url, client_registration=registration))
 
         return result, VulnFlag(False)
 
@@ -87,10 +55,10 @@ class ClientScanner(Scanner[Clients], Need2[Realms, WellKnownDict]):
         except ValueError as e:
             super().warn(f'{url} is not json but status is 200. It\'s weird. {e}')
 
-    def has_auth_endpoint(self, client_name: str, realm: Realm, well_known_dict: WellKnownDict) -> str:
+    def has_auth_endpoint(self, client_name: str, realm: Realm, well_known: WellKnown) -> str:
         try:
 
-            auth_url = well_known_dict[realm.name].json['authorization_endpoint']
+            auth_url = well_known.json['authorization_endpoint']
 
             r = super().session().get(auth_url, params={'client_id': client_name}, allow_redirects=False)
             # TODO : is code 400 always an existing client ?
@@ -101,7 +69,7 @@ class ClientScanner(Scanner[Clients], Need2[Realms, WellKnownDict]):
 
         except KeyError as e:
             print(
-                f'realm {realm.name}\'s wellknown doesn\'t exists or do not have "authorization_endpoint". ({well_known_dict})')
+                f'realm {realm.name}\'s wellknown doesn\'t exists or do not have "authorization_endpoint". ({well_known})')
             print(e)
             auth_url = None
         return auth_url
